fix types definition & add addresses functions to utils
